requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2
django-cors-headers==4.4.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fcm-django==2.1.0
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2